+++ CHANGELOG.md
@@ -5,15 +5,22 @@
 - [Server] Read only replicas and high availability with auto roll-over and
   promotion (Raft consensus based)
   [#72](https://github.com/datalevin/datalevin/issues/72). This implementation
-  comes with a Jepsen test suite for checking correctness
+  comes with a Jepsen test suite to validate correctness
   [#252](https://github.com/datalevin/datalevin/issues/252). Details in
   [doc](doc/ha.md).
 - [API] Java API. Release Datalevin as a Java library to maven central.
 - [API] JSON API. [#182](https://github.com/datalevin/datalevin/issues/182)
 - [MCP] Built-in stdio based MCP server. Details in [doc](doc/mcp.md).
 - [Datalog] allow to register a `:db/udf` to specify a user defined function.
-  This is cross platform, will resolve in runtime environment and does not
-  restore in DB.
+  This is cross language, resolves in runtime environment and does not
+  persist in DB.
+- [Datalog] allow `:db/embedding true` property for string attributes, which
+  will use an embedding provider to embed the texts into vectors and index them.
+  A trimmed-down build of llama.cppA is included as the default embedding
+  provider. The default embedding model `multilingual-e5-small-Q8_0.gguf` is
+  downloaded from HuggingFace on first use.
+- [Datalog] for embedding enabled datoms, `embedding-neighbors` built-in function to
+  return `[e, a, v]` based on vector similarity. Details in [doc](doc/vector.md).
 
 ## 0.10.7 (2026-03-03)